requirements.txt
faiss-cpu
numpy
pandas
sentence-transformers
streamlit
